relocate mock user
